backend/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
billiard==4.2.1
black==24.10.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.3
flower==2.0.1
h11==0.14.0
httptools==0.6.4
humanize==4.11.0
idna==3.10
kombu==5.4.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
peewee==3.17.7
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
